[PP-47] splitting users components
